add Validation in PostController

+++ src/main/java/com/pongchi/glimelight/api/v1/PostController.java
@@ -1,9 +1,13 @@
 package com.pongchi.glimelight.api.v1;
 
+import static com.pongchi.glimelight.api.v1.dto.ResponseDto.createResponseEntity;
+import static com.pongchi.glimelight.api.v1.dto.ResponsesDto.createResponsesEntity;
+
 import java.util.UUID;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,8 +15,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import static com.pongchi.glimelight.api.v1.dto.ResponseDto.createResponseEntity;
-import static com.pongchi.glimelight.api.v1.dto.ResponsesDto.createResponsesEntity;
 import com.pongchi.glimelight.api.v1.dto.post.PostCreateRequestDto;
 import com.pongchi.glimelight.common.ResponseCode;
 import com.pongchi.glimelight.service.PostService;
@@ -26,7 +28,14 @@ public class PostController {
     private final PostService postService;
 
     @PostMapping("/api/v1/posts")
-    public ResponseEntity<?> create(@RequestBody PostCreateRequestDto requestDto) {
+    public ResponseEntity<?> create(@RequestBody PostCreateRequestDto requestDto, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return createResponsesEntity(
+                ResponseCode.INVALID_PARAMETER, 
+                bindingResult.getAllErrors()
+            );
+        }
+
         return createResponseEntity(
             ResponseCode.SUCCESS,
             postService.create(requestDto)
@@ -34,7 +43,14 @@ public class PostController {
     }
 
     @GetMapping("/api/v1/posts")
-    public ResponseEntity<?> postListAll(Pageable pageable) {
+    public ResponseEntity<?> postListAll(Pageable pageable, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return createResponsesEntity(
+                ResponseCode.INVALID_PARAMETER, 
+                bindingResult.getAllErrors()
+            );
+        }
+
         return createResponsesEntity(
             ResponseCode.SUCCESS,
             postService.findAll(pageable).getContent()
@@ -42,7 +58,14 @@ public class PostController {
     }
 
     @GetMapping("/api/v1/posts/{id}")
-    public ResponseEntity<?> read(@PathVariable("id") UUID id) {
+    public ResponseEntity<?> read(@PathVariable("id") UUID id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return createResponsesEntity(
+                ResponseCode.INVALID_PARAMETER, 
+                bindingResult.getAllErrors()
+            );
+        }
+
         return createResponseEntity(
             ResponseCode.SUCCESS,
             postService.read(id)
@@ -50,7 +73,14 @@ public class PostController {
     }
 
     @DeleteMapping("/api/v1/posts/{id}")
-    public ResponseEntity<?> delete(@PathVariable("id") UUID id) {
+    public ResponseEntity<?> delete(@PathVariable("id") UUID id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return createResponsesEntity(
+                ResponseCode.INVALID_PARAMETER, 
+                bindingResult.getAllErrors()
+            );
+        }
+        
         return createResponseEntity(
             ResponseCode.SUCCESS,
             postService.delete(id)
